Organize component subfolders
